upgrade-toolkit/CDH-Discovery-Tool/requirements.txt
cm-client==33.0.0
requests==2.27.1
PyYAML==6.0
jaydebeapi==1.2.3
openpyxl==3.0.9
tzlocal==4.2
pandas==1.1.5